requirements.txt
python-telegram-bot
transformers
torch
torchvision
Pillow
speechrecognition
whisper
Flask
